v2/requirements.txt
google-generativeai
numpy
polars
pyarrow
pypdf
python-dotenv
streamlit